requirements.txt
# NeuroCore - Perceptron Analytics Dashboard
# Requirements for Streamlit Application

streamlit>=1.28.0
streamlit-lottie
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
liac-arff>=2.5.0
requests>=2.31.0
